refactor(canvas): move session activity projections
