+++ apps/canvas-workspace/src/renderer/src/agent-chat/sessions/useScopeRunningSessions.test.tsx
@@ -3,12 +3,12 @@ import { act } from 'react';
 import { createRoot, type Root } from 'react-dom/client';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 import { useScopeRunningSessions } from './useScopeRunningSessions';
-import { conversationKey } from '../../../../../shared/conversation-runtime';
+import { conversationKey } from '../../../../shared/conversation-runtime';
 import {
   resetConversationStoreForTests,
   setConversationLoading,
   setConversationMessages,
-} from '../../../agent-chat/runtime/conversationStore';
+} from '../runtime/conversationStore';
 
 (globalThis as { IS_REACT_ACT_ENVIRONMENT?: boolean }).IS_REACT_ACT_ENVIRONMENT = true;
